+++ sw/simulation/agent.cpp
@@ -5,7 +5,7 @@ Agent::Agent()
   activated = false;
 }
 
-float Agent::get_position(uint8_t dim)
+float Agent::get_position(uint16_t dim)
 {
   if (dim < 3) {
     return state[dim];
@@ -18,7 +18,7 @@ float Agent::get_orientation()
   return state[6];
 }
 
-float Agent::get_state(const uint8_t i)
+float Agent::get_state(const uint16_t i)
 {
   float noise = rg.gaussian_float(0.0, NOISE_R);
   return state[i] + noise;
