+++ sw/simulation/controllers/aggregation/controller_aggregation.h
@@ -28,7 +28,7 @@ public:
   /**
    * Implementation of method to get the commanded velocity
    */
-  virtual void get_velocity_command(const uint8_t ID, float &v_x, float &v_y);
+  virtual void get_velocity_command(const uint16_t ID, float &v_x, float &v_y);
 };
 
 #endif /*CONTROLLER_AGGREGATION_H*/
