+++ sw/simulation/controllers/behavior_tree/behavior_tree_wheeled.cpp
@@ -33,7 +33,7 @@ behavior_tree_wheeled::behavior_tree_wheeled() : Controller()
   BLKB.set("wheelSpeed1", 0.); // Output 1
 }
 
-void behavior_tree_wheeled::get_velocity_command(const uint8_t ID, float &v_x, float &v_y)
+void behavior_tree_wheeled::get_velocity_command(const uint16_t ID, float &v_x, float &v_y)
 {
   v_x = 0;
   v_y = 0;
