+++ sw/simulation/controllers/behavior_tree/behavior_tree_wheeled.h
@@ -39,7 +39,7 @@ public:
   /**
    * Update the inputs and run the behavior tree command
    */
-  virtual void get_velocity_command(const uint8_t ID, float &v_x, float &v_y);
+  virtual void get_velocity_command(const uint16_t ID, float &v_x, float &v_y);
 };
 
 #endif /*BEHAVIOR_TREE_WHEELED_H*/
