+++ sw/simulation/controllers/pattern_formation/controller_pattern.cpp
@@ -18,7 +18,7 @@ controller_pattern::controller_pattern() : controller_lattice_basic()
   beta_des.push_back(M_PI / 2.0 + atan(_ddes_x / _ddes_y));
 }
 
-void controller_pattern::get_velocity_command(const uint8_t ID, float &v_x, float &v_y)
+void controller_pattern::get_velocity_command(const uint16_t ID, float &v_x, float &v_y)
 {
   v_x = 0.0;
   v_y = 0.0;
