+++ sw/simulation/controllers/pattern_formation/controller_pattern.h
@@ -22,7 +22,7 @@ class controller_pattern : public controller_lattice_basic
 public:
   controller_pattern();
   ~controller_pattern() {};
-  virtual void get_velocity_command(const uint8_t ID, float &v_x, float &v_y);
+  virtual void get_velocity_command(const uint16_t ID, float &v_x, float &v_y);
 };
 
 #endif /*CONTROLLER_PATTERN*/
