+++ sw/simulation/controllers/random_exploration/random_exploration.cpp
@@ -7,7 +7,7 @@ random_exploration::random_exploration() : Controller()
   moving = false;
 }
 
-void random_exploration::get_velocity_command(const uint8_t ID, float &v_x, float &v_y)
+void random_exploration::get_velocity_command(const uint16_t ID, float &v_x, float &v_y)
 {
   v_x = 0.0;
   v_y = 0.0;
